[REF] odootil: _wrap_empty_iterable
When we check for CRUD access, object can be empty (Model), but we still
must iterate at least once in that case. So we wrap empty object with a
list.

[ADD] odootil: new_with_recs

Method to imitate record update with values without actually writing
changes to database. This is wrapper for `new` method, which only
creates new virtual record with specified values only, but can't use
existing record/records as a base.